Rename projection variables
It looks like you got these variable names from example code that
proj4js has in their documentation. Those names are useful for them to
show how the order of the arguments affects the output of their
projection function, but they do little to help developers reading our
code understand why or what it is that we're doing.