Change `let`s to `const`s where appropriate
Conceptually, these values should never change. The ESRI projection
will always be the ESRI projection, and the WSGI84 projection will
always be the WSGI84 projection. Because of that, it's better to use
`const` for these values, because it tells other programmers reading
our code that these are constants, and if anyone tries to reassign
these values, the JavaScript interpreter will yell at them.