Move `projCoord` variable closer to where it's used
Since `projCoord` is only being used within this loop, it can be
confusing to see it declared far away from there. There's also no
reason for it to be set to 0 initially; it's a coordinate, not a
number, and we would never want to accidentally try to use it like a
coordinate when it was a number.