Remove implicit dependency on `data.json` file
If another developer wanted to use this package to convert their own
GeoJSON data, they would have to move the file into the package folder
and rename it to `data.json`, then they would have to start the
`webpack-dev-server`, open a browser, and copy-and-paste the value
displayed there. By removing the import that uses `data.json` and
instead making a function to which this data will be passed (in the
form of the parameter `featureCollection`), this code becomes far more
versatile and easier to use. Now, there doesn't even have to be a
file--if they generated some coordinates somehow, they could simply
pass them to our function instead of saving that output to a file and
going through the other process. If they wanted these new coordinates
saved to disk, they would have the option of doing so in whatever
manner worked for them, be it printing them to the console and copying
them or using Node's filesystem functions to send it directly to a file.e