Change `import` to `require`
Though `import` is nice, and I advocate using it in projects that use
Webpack, Node does not yet support this syntax for some complex
reasons. So, by changing this to `require`, we get essentially the
same functionality and Webpack is just as capable of parsing our
dependencies later should we or someone else want to use this.